+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Make document title configurable
 
 ## 1.0.0-beta (2018-04-03)